Moved getVehicleOptions to service
